Changed bit declarations from unsigned int to byte (unsigned char). This is more space efficient, but was really done because there is some issue with MSVC when the unsigned int is used.  There must be some alignment and/or displacement kind of problem with the unsigned int. I don't have the time/energy/interest to try to track down the exact reason....sorry.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8480 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/tab_structs.h
@@ -629,10 +629,10 @@ enum SubgoalFrameType {
 typedef struct subgoal_frame {
   byte sf_type;		  /* The type of subgoal frame */
   byte is_complete;	  /* If producer, whether its answer set is complete */
-  unsigned int  is_reclaimed:1;	  /* Whether structs for supporting answer res from an
+  byte is_reclaimed:1;	  /* Whether structs for supporting answer res from an
 			     incomplete table have been reclaimed */
-  unsigned int negative_initial_call:1;
-  unsigned int unused:6;
+  byte negative_initial_call:1;
+  byte unused:6;
   byte visited;
   TIFptr tif_ptr;	  /* Table of which this call is a part */
   BTNptr leaf_ptr;	  /* Handle for call in the CallTrie */
